requirements.txt
python-frontmatter>=1.1.0
rich>=13.0.0
scikit-learn>=1.3.0
